requirements.txt
asgiref==3.7.2
#backports.zoneinfo==0.2.1;python_version<"3.9"
Django==4.2.3
django-cors-headers==4.2.0
djangorestframework==3.14.0
pytz==2023.3
sqlparse==0.4.4
#typing-extensions==4.7.1
gunicorn==20.1.0
django_simple_history==3.3.0
whitenoise==6.5.0
djangorestframework-simplejwt==5.2.2
#psycopg-binary>=3.1.10
psycopg2-binary==2.9.9
s3transfer==0.6.1
boto3==1.26.135
botocore==1.29.135
django-storages==1.13.2
secure-smtplib==0.1.1
